GUI Update + More
Updated GUI to include life count and static function to remove number a
life. Also consolidated spriteSheet with ImageLoader. Changed the size
of board pieces and created constants for the size of the GUI panels so
that if you change the size of that constant the GUI will match it. Also
moved pacSheet.png into the same package as everything else.